refactor: Member 패키지, 엔티티 변경에 따른 충돌 해결
