+++ backend/main-service/src/main/java/kkakka/mainservice/cart/application/CartService.java
@@ -7,6 +7,8 @@ import kkakka.mainservice.cart.domain.repository.CartRepository;
 import kkakka.mainservice.cart.ui.dto.CartRequestDto;
 import kkakka.mainservice.member.domain.Member;
 import kkakka.mainservice.member.domain.repository.MemberRepository;
+import kkakka.mainservice.member.member.domain.Member;
+import kkakka.mainservice.member.member.domain.repository.MemberRepository;
 import kkakka.mainservice.product.domain.Product;
 import kkakka.mainservice.product.domain.repository.ProductRepository;
 import org.springframework.stereotype.Service;
