+++ backend/main-service/src/main/java/kkakka/mainservice/cart/domain/Cart.java
@@ -1,14 +1,19 @@
 package kkakka.mainservice.cart.domain;
 
-import kkakka.mainservice.member.domain.Member;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
+import kkakka.mainservice.member.member.domain.Member;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Table(name = "cart")
