--- backend/main-service/src/main/java/kkakka/mainservice/member/domain/repository/MemberRepository.java
@@ -1,7 +1,0 @@
-package kkakka.mainservice.member.domain.repository;
-
-import kkakka.mainservice.member.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-}
